Added more tags and a couple advanced tags that are for expert users only.
gedcomNews:limit:flag
totalGedcomNews
callBlock:block_name:config=value:config2=value
includeFile:path/filename

callBlock will output the contents of the block.  Just please don't call the html_plus/Advanced HTML block with this, the results would be scary :)  This tag is currently very flaky with some functions.  For example the random media block will not work correctly but the Gallery2 one works fine.

Example:
    #callBlock:recent_changes:cache=1:days=10:hide_empty=no#

includeFile will include the file just as if an include statement was called.  But, for security reasons it can nly be used on the GEDCOM welcome page, not the user portals.  Seriously, you do not want the public adding this tag to their portals :)


git-svn-id: svn://svn.webtrees.net/trunk@3236 4201c862-f99b-1943-94d0-25fd05010cbc